+++ Changelog.md
@@ -1,6 +1,8 @@
 # Change log
 
 ## master (unreleased)
+### New features
+* Parsing `OleObject` in `GraphicFrame`
 
 ## 0.2.0 (2017-03-25)
 ### New features